vegan_ecommerce/requirements/base.txt
# Dev environment
pip

# App
Django==5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
django-environ==0.11.2
psycopg2-binary==2.9.7
python-dateutil==2.9.0.post0
django-filter==24.3
django-cors-headers==4.5.0
djoser==2.3.1
drf-nested-routers==0.94.1

# swagger
drf-spectacular==0.27.2